Add files via upload
This is a code from a computer lab session where we first practically applied reading/importing datasets and analysing them. We analysed the exchange rates of different currencies compared to one another and plotted the results. We extracted the data from the Federal Reserve Economic Database. I was not able to complete part 2.d at the time but I am going to comeback to it soon to attempt again.